fix weird state handling by adding custom hook
window.removeEventListener was not removing the callback function. removeEventListener(callback) only removes the callback functions, that are equal to the one given in addEventListener(callback). But because the callback function was defined in scope of a functionalal compoent, it was redefined each new render.

Dan Ambramov describes a similar bug with setInterval(delay, cb) and clearInterval(cb) and solves it with an own hook. See https://overreacted.io/making-setinterval-declarative-with-react-hooks/